backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
starlette==0.27.0  # Compatible with FastAPI 0.104.1
anyio>=3.7.1,<4.0.0  # Compatible with FastAPI 0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.0

# LLM Provider Dependencies
openai>=1.0.0
anthropic>=0.7.0

# Search and Data Processing
pandas>=2.0.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Email Support
sendgrid>=6.10.0

# Day 2 Browser Automation Dependencies
# Note: browser-use is not available on PyPI, using playwright directly
playwright>=1.40.0
playwright-stealth>=1.0.6
apify-client>=1.0.0
asyncio-throttle>=1.0.2
urllib3>=1.26.0
typing-extensions>=4.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Optional but recommended
aiohttp>=3.8.0

# Day 3 Browser Streaming Dependencies
docker>=6.1.0
websockets>=11.0.0
browser-use>=0.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0